Initial Version
This is the Legacy version of Insane RT Framework, it only compile well in MSVC QT kits , so if you want make it work in other platforms , changes must be made on qmake scripts. The new version will use only MinGW , and be prepeared to multiplatform compilations without changes. Good Luck! :)